pipeline/requirements.txt
# ============================================
# GenHack Climate - Python Dependencies
# ============================================
# Phase 1: Core geospatial stack + ML prep

# Geospatial core
rasterio>=1.3.9
rioxarray>=0.15.0
xarray>=2023.1.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.5

# Array computation
numpy>=1.24.0
pandas>=2.0.0

# Validation & contracts
pydantic>=2.5.0

# Templating & reporting
jinja2>=3.1.2
weasyprint>=60.0

# Plotting
matplotlib>=3.7.0

# CLI
click>=8.1.0
pyyaml>=6.0.1

# Cloud integrations
google-cloud-storage>=2.10.0
google-cloud-logging>=3.8.0

# Climate data (Phase 2+)
cdsapi>=0.6.1

# Optional: Future ML dependencies (Phase 2+)
# torch>=2.0.0
# torchvision>=0.15.0
# scikit-learn>=1.3.0
